+++ apps/web/src/features/notification/ui/index.ts
@@ -1,1 +1,1 @@
-export { NotificationReadStatusIndicator } from './notification-read-status-indicator';
+export { NotificationListItem } from './notification-list-item';
