+++ apps/web/src/features/notification/ui/notification-list-item/NotificationListItem.tsx
@@ -1,15 +1,17 @@
 'use client';
 
-import { Button, HStack } from '@causw/cds';
-
-import { NotificationReadStatusIndicator } from '@/features/notification';
+import { Button, HStack, mergeStyles } from '@causw/cds';
 
 import {
+  NOTIFICATION_TYPE,
+  NotificationReadStatusIndicator,
   NotificationTimeLabel,
   NotificationTitle,
   NotificationTypeLabel,
   type GetNotificationsResponseDto,
 } from '@/entities/notification';
+
+import { useNotificationListItem } from '../../model';
 
 interface NotificationListItemProps {
   notification: GetNotificationsResponseDto;
@@ -18,8 +20,18 @@ interface NotificationListItemProps {
 export const NotificationListItem = ({
   notification,
 }: NotificationListItemProps) => {
+  const { handleClickNotification } = useNotificationListItem();
+
   return (
-    <Button className="relative flex h-fit w-full flex-col gap-2 rounded-lg bg-white px-4 py-4.5">
+    <Button
+      className={mergeStyles(
+        'relative flex h-fit w-full flex-col gap-2 rounded-lg bg-white px-4 py-4.5',
+        notification.isRead &&
+          notification.noticeType === NOTIFICATION_TYPE.SYSTEM.type &&
+          'hover:bg-white!',
+      )}
+      onClick={() => handleClickNotification(notification)}
+    >
       <NotificationReadStatusIndicator isRead={notification.isRead} />
       <HStack className="w-full items-center justify-between">
         <NotificationTypeLabel noticeType={notification.noticeType} />
